+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 2025.01.27
+- performance improvements
+- only register updates when stats are requested by an addon
 - update ult cost periodically to account for some situations where a debuff affects the ult cost of a player ( ex. Cloudrest excecute )
 
 ## 2025.01.23